move case_aggregates_by_month
